Add release script
